Add ContentOnly option for directory transfers
Normally, copying a directory also creates a directory with the same
name on the target side of the transfer. For example:

    $ scp -r source remote:/some/path
    source/file1           > remote:/some/path/source/file1
    source/folder1/file2   > remote:/some/path/source/folder1/file2

If we don't want to create a new directory on the target side we can use
a glob after the source, indicating an scp operation for each file or
folder inside:

    $ scp -r source/* remote:/some/path
    source/file1           > remote:/some/path/file1
    source/folder1/file2   > remote:/some/path/folder1/file2

When ContentOnly is true, creation of the source directory on the target
side is skipped and the contents are copied into the existing directory
instead, using a single connection and job queue.